requirements.txt
# Core LangChain and AI
langchain>=0.1.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-chroma>=0.1.0
openai>=1.0.0

# Web scraping and HTTP
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
scholarly>=1.7.11
arxiv>=1.4.8

# PDF processing and OCR
pytesseract>=0.3.10
pdf2image>=1.16.3
PyPDF2>=3.0.1
pdfplumber>=0.10.3
Pillow>=10.0.0

# Vector database and embeddings
chromadb>=0.4.0
numpy>=1.24.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0

# Frontend
streamlit>=1.28.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0